Sidecar: use env var for the claim file folder.
Removed hardcoded folder where the claim file should be read.
Also, the config/samples folder has been modified so a sample run can be
triggered with:
oc kustomize config/samples | oc apply -f -

The kustomization file contains the cnf suite configmap, the preflight
secret and a CnfCertificationSuiteRun CR to trigger the test suite using
the tnf_config.yaml and the dockerconfig file from the configMap and the
secret.
All the resources will be created in the "cnf-certification-operator"
namespace.