RBAC changes and test bundle for installation via OLM
The supported installModes have been changed to ownNamespace only.

This commit separates the cluster permissions in two different services
accounts:
- manager/controller SA: changed from ClusterRole/ClusterRoleBinding to
  Role/RoleBinding, so the controller has the minimum required
  permissions to deal with Run/Report CRs.
- cnf certsuite pod SA: a cluster-wide permissions with
  ClusterRole/ClusterRoleBinding that grants full access to every
  resource in the cluster.

Also, some changes were done in the config/* structure in order to be able
to deploy and run the operator using the "make deploy" command.

An initial bundle for OLM was created using the "make bundle" command.
In order to test it, both a bundle and a catalog need to be built using
"make" commands plus the appropriated env vars as defined in the
operator-sdk docs:
- make bundle-build bundle-push
- make catalog-build catalog-push
See:
- https://sdk.operatorframework.io/docs/olm-integration/quickstart-bundle/#steps
- https://sdk.operatorframework.io/docs/olm-integration/tutorial-bundle/#deploying-bundles-in-production