Set console plugin ns dynamically (#100)
Based on
[this](https://github.com/test-network-function/cnf-certsuite-operator/pull/89#discussion_r1681039870)
discussion in previous PR, Instead of using a hard coded ns for the
console plugin, the plugin ns will be set according to the ns of the
controller.

The controller ns is saved as an env var within the container using
[downward
API](https://github.com/kubernetes/kubernetes/blob/release-1.0/docs/user-guide/downward-api.md).
That way, when setting up controller's manager and creating the plugin
resources, we can extract the controller's ns and set it as the console
plugin ns as well.

The plugin resources don't have a default ns, and are set with a place
holder for the ns. That way, if we fail to set the operator ns as the
plugins ns, will get back an error.